% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{rx_media}
\alias{rx_media}
\title{Extract the media tibble from a relational result.}
\usage{
rx_media(x)
}
\arguments{
\item{x}{An \code{rx_relational} object (or any object).}
}
\value{
A tibble with media columns when \code{x} is an \code{rx_relational},
otherwise \code{tibble::tibble()}.
}
\description{
Extract the media tibble from a relational result.
}
