% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{rx_users}
\alias{rx_users}
\title{Extract the users tibble from a relational result.}
\usage{
rx_users(x)
}
\arguments{
\item{x}{An \code{rx_relational} object (or any object).}
}
\value{
A tibble with user columns when \code{x} is an \code{rx_relational},
otherwise \code{tibble::tibble()}.
}
\description{
Extract the users tibble from a relational result.
}
